[4][eva] feature: modify SuperSmartParkingBoyTest to adapt park result as ParkResult

+++ src/main/java/com/oocl/cultivation/SuperSmartParkingBoy.java
@@ -2,7 +2,7 @@ package com.oocl.cultivation;
 
 import java.util.List;
 
-public class SuperSmartParkingBoy extends SmartParkingBoy{
+public class SuperSmartParkingBoy extends SmartParkingBoy {
 
     private final List<ParkingLot> parkingLots;
 
@@ -13,9 +13,9 @@ public class SuperSmartParkingBoy extends SmartParkingBoy{
     @Override
     public ParkResult park(Car car) {
         ParkingLot maxPositionParkingLot = parkingLots.get(0);
-        for (ParkingLot parkingLot: parkingLots) {
-            double availableRate = parkingLot.getEmptyPosition() / parkingLot.getCapacity();
-            double maxAvailableRate = maxPositionParkingLot.getEmptyPosition() / maxPositionParkingLot.getCapacity();
+        for (ParkingLot parkingLot : parkingLots) {
+            double availableRate = (double) parkingLot.getEmptyPosition() / parkingLot.getCapacity();
+            double maxAvailableRate = (double) maxPositionParkingLot.getEmptyPosition() / maxPositionParkingLot.getCapacity();
             if (Double.compare(availableRate, maxAvailableRate) > 0) {
                 maxPositionParkingLot = parkingLot;
             }
